Safari-609.1.11.1/LayoutTests/media/video-seek-double-expected.txt

Test that seeking twice in the same run loop will cancel the first seek.


EVENT(canplaythrough)
RUN(video.currentTime = 2)
RUN(video.currentTime = 0.5)

EVENT(seeked)
EXPECTED (video.currentTime == '0.5') OK
RUN(video.play())
EVENT(play)
END OF TEST
